[FIX] team > owner is OneToOne
